Add support for opening a connection from an existing stream
Rework of an earlier PR, so that a stdlib stream can be placed in
LdapConnSettings and used to initialize the connection. To keep
the settings cloneable, the stream-bearing enum is cloned to an
invalid variant, since the streams themselves can't be cloned.

The provided stream is automatically set to nonblocking mode.